working client communication
